+++ main.go
@@ -21,7 +21,10 @@ import (
 	"github.com/steadybit/extension-kit/exthttp"
 	"github.com/steadybit/extension-kit/extlogging"
 	"github.com/steadybit/extension-kit/extruntime"
-	_ "net/http/pprof" //allow pprof
+	_ "go.uber.org/automaxprocs" // Importing automaxprocs automatically adjusts GOMAXPROCS.
+	// You can find more details of its behavior from the doc comment of memlimit.SetGoMemLimitWithEnv.
+	_ "github.com/KimMachineGun/automemlimit" // By default, it sets `GOMEMLIMIT` to 90% of cgroup's memory limit.
+	_ "net/http/pprof"           //allow pprof
 )
 
 func main() {